^C while waiting for jupyter runs qdel
Also, the -i JOBID option now works.

Not that trapping the ^C was a little tricky -- it required using "ssh -t" to
launch the job, but that ended up adding carriage returns (\r) in the log file.
So when pulling things like the jupyter port number out of log file, we need to
translate '\r' to ' ' (since we eventually delete all spaces from the
variables).

Also, I condensed all the ssh calls to a single line (each) rather than using
HEREDOCs.